+++ src/renderer/page/subscriptions/index.js
@@ -1,4 +1,5 @@
 import { connect } from 'react-redux';
+import * as settings from 'constants/settings';
 import {
   selectSubscriptionClaims,
   selectSubscriptions,
@@ -6,25 +7,26 @@ import {
   selectIsFetchingSubscriptions,
   selectNotifications,
 } from 'redux/selectors/subscriptions';
-import { doFetchClaimsByChannel } from 'redux/actions/content';
 import { setSubscriptionNotifications, doFetchMySubscriptions } from 'redux/actions/subscriptions';
+import { doSetClientSetting } from 'redux/actions/settings';
+import { makeSelectClientSetting } from 'redux/selectors/settings';
 import SubscriptionsPage from './view';
 
 const select = state => ({
   loading:
     selectIsFetchingSubscriptions(state) ||
     Boolean(Object.keys(selectSubscriptionsBeingFetched(state)).length),
-  subscriptionsBeingFetched: selectSubscriptionsBeingFetched(state),
   subscriptions: selectSubscriptions(state),
   subscriptionClaims: selectSubscriptionClaims(state),
   notifications: selectNotifications(state),
+  autoDownload: makeSelectClientSetting(settings.AUTO_DOWNLOAD)(state),
 });
 
 export default connect(
   select,
   {
-    doFetchClaimsByChannel,
     setSubscriptionNotifications,
     doFetchMySubscriptions,
+    doSetClientSetting,
   }
 )(SubscriptionsPage);