+++ src/renderer/page/subscriptions/view.jsx
@@ -1,26 +1,32 @@
 // @flow
 import React from 'react';
 import Page from 'component/page';
+import * as settings from 'constants/settings';
 import type { Subscription } from 'types/subscription';
 import * as NOTIFICATION_TYPES from 'constants/notification_types';
 import Button from 'component/button';
 import FileList from 'component/fileList';
 import type { Claim } from 'types/claim';
-import isDev from 'electron-is-dev';
 import HiddenNsfwClaims from 'component/hiddenNsfwClaims';
+import { FormField, FormRow } from 'component/common/form';
 
 type Props = {
-  doFetchClaimsByChannel: (string, number) => void,
   doFetchMySubscriptions: () => void,
   setSubscriptionNotifications: ({}) => void,
   subscriptions: Array<Subscription>,
   subscriptionClaims: Array<{ uri: string, claims: Array<Claim> }>,
-  subscriptionsBeingFetched: {},
   notifications: {},
   loading: boolean,
+  autoDownload: boolean,
+  doSetClientSetting: (string, boolean) => void,
 };
 
 export default class extends React.PureComponent<Props> {
+  constructor() {
+    super();
+    (this: any).onAutoDownloadChange = this.onAutoDownloadChange.bind(this);
+  }
+
   componentDidMount() {
     const { notifications, setSubscriptionNotifications, doFetchMySubscriptions } = this.props;
     doFetchMySubscriptions();
@@ -37,43 +43,12 @@ export default class extends React.PureComponent<Props> {
     setSubscriptionNotifications(newNotifications);
   }
 
-  componentDidUpdate() {
-    const {
-      subscriptions,
-      subscriptionClaims,
-      doFetchClaimsByChannel,
-      subscriptionsBeingFetched,
-    } = this.props;
-
-    const subscriptionClaimMap = {};
-    subscriptionClaims.forEach(claim => {
-      /*
-      This check added 6/20/18 to fix function receiving empty claims unexpectedly.
-      The better fix is ensuring channels aren't added to byId if there are no associated claims
-      We are adding this now with the redesign release to ensure users see the correct subscriptions
-      */
-      if (claim.claims.length) {
-        subscriptionClaimMap[claim.uri] = 1;
-      } else if (isDev) {
-        // eslint-disable no-console
-        console.error(
-          `Claim for ${
-            claim.uri
-          } was added to byId in redux but there are no loaded fetched claims. This shouldn't happen because a subscription should have claims attached to it.`
-        );
-        // eslint-enable no-console
-      }
-    });
-
-    subscriptions.forEach(sub => {
-      if (!subscriptionClaimMap[sub.uri] && !subscriptionsBeingFetched[sub.uri]) {
-        doFetchClaimsByChannel(sub.uri, 1);
-      }
-    });
+  onAutoDownloadChange(event: SyntheticInputEvent<*>) {
+    this.props.doSetClientSetting(settings.AUTO_DOWNLOAD, event.target.checked);
   }
 
   render() {
-    const { subscriptions, subscriptionClaims, loading } = this.props;
+    const { subscriptions, subscriptionClaims, loading, autoDownload } = this.props;
 
     let claimList = [];
     subscriptionClaims.forEach(claimData => {
@@ -85,6 +60,15 @@ export default class extends React.PureComponent<Props> {
     return (
       <Page notContained loading={loading}>
         <HiddenNsfwClaims uris={subscriptionUris} />
+        <FormRow alignRight>
+          <FormField
+            type="checkbox"
+            name="auto_download"
+            onChange={this.onAutoDownloadChange}
+            checked={autoDownload}
+            prefix={__('Automatically download new content from your subscriptions')}
+          />
+        </FormRow>
         {!subscriptions.length && (
           <div className="page__empty">
             {__("It looks like you aren't subscribed to any channels yet.")}